+++ src/CLI.hs
@@ -15,6 +15,8 @@ import Lentille.GitHub.Organization qualified as GH_ORG
 import Lentille.GitHub.PullRequests qualified as GH_PR
 import Lentille.GitHub.UserPullRequests qualified as GH_UPR
 import Lentille.GitHub.Watching qualified
+import Lentille.GitLab.Group qualified as GL_GROUP
+import Lentille.GitLab.MergeRequests qualified as GL_MR
 import Lentille.GraphQL (newGraphClient)
 import Lentille.Jira qualified as Jira
 import Macroscope.Main (runMacroscope)
@@ -206,6 +208,8 @@ usageLentille =
         <> mkSubCommand "github-changes" "Get changes list" githubChangesUsage
         <> mkSubCommand "github-user-changes" "Get changes list" githubUserChangesUsage
         <> mkSubCommand "github-watching" "Get watched list" githubWatchingUsage
+        <> mkSubCommand "gitlab-projects" "Get projects list" gitlabProjectsUsage
+        <> mkSubCommand "gitlab-changes" "Get changes list" gitlabChangesUsage
         <> mkSubCommand "jira-issues" "Get recent issues" jiraIssuesUsage
     )
  where
@@ -274,6 +278,30 @@ usageLentille =
     io (url, secret, user, since, limitM, humanF) = runStandaloneStream do
       client <- getGraphClient url secret
       let stream = GH_UPR.streamUserPullRequests client (const Nothing) (toSince since) user
+      if humanF
+        then dumpChanges formatChange limitM stream
+        else dumpJSON limitM stream
+
+  gitlabProjectsUsage = io <$> parser
+   where
+    parser = (,,) <$> urlOption <*> secretOption <*> orgOption
+    io (url, secret, org) = runStandaloneStream do
+      client <- getGraphClient url secret
+      dumpJSON Nothing $ GL_GROUP.streamGroupProjects client org
+
+  gitlabChangesUsage = io <$> parser
+   where
+    parser =
+      (,,,,,)
+        <$> urlOption
+        <*> secretOption
+        <*> projectOption
+        <*> sinceOption
+        <*> limitOption
+        <*> humanOption
+    io (url, secret, repo, since, limitM, humanF) = runStandaloneStream do
+      client <- getGraphClient url secret
+      let stream = GL_MR.streamMergeRequests client (const Nothing) (toSince since) repo
       if humanF
         then dumpChanges formatChange limitM stream
         else dumpJSON limitM stream